+++ src/main/java/cc/cassian/raspberry/blocks/AshBlock.java
@@ -1,0 +1,39 @@
+package cc.cassian.raspberry.blocks;
+
+import cc.cassian.raspberry.registry.RaspberryBlocks;
+import net.mehvahdjukaar.supplementaries.common.block.blocks.AshLayerBlock;
+import net.mehvahdjukaar.supplementaries.common.entities.FallingAshEntity;
+import net.mehvahdjukaar.supplementaries.reg.ModRegistry;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.item.FallingBlockEntity;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Material;
+
+public class AshBlock extends AshLayerBlock {
+    private static final Block ASH_LAYER = ModRegistry.ASH_BLOCK.get();
+
+    public AshBlock(Properties properties) {
+        super(properties);
+        this.registerDefaultState(this.stateDefinition.any().setValue(LAYERS, 8));
+
+    }
+
+    @Override
+    public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
+        if (isFree(level.getBlockState(pos.below())) && pos.getY() >= level.getMinBuildHeight()) {
+            FallingBlockEntity fallingBlockEntity = FallingBlockEntity.fall(level, pos, state);
+            this.falling(fallingBlockEntity);
+        }
+    }
+
+
+    public static boolean isFree(BlockState state) {
+        Material material = state.getMaterial();
+        if (state.getBlock() instanceof AshBlock || state.getBlock() instanceof GravitationalSnowLayerBlock) return false;
+        return state.isAir() || state.is(BlockTags.FIRE) || material.isLiquid() || material.isReplaceable();
+    }
+}