+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Storage block for Ash (Supplementaries).
+
 ### Fixed
 - Foxes no longer drop Rabbit's Feet.
 - Launching no longer shows as an option for axes in EMI.